Check if file exists before deleting it when rebuilding classpath in JavaComplete2

+++ rplugin/python3/VimStudio/Configurator.py
@@ -17,7 +17,8 @@ class Configurator(object):
     def resetJavacomplete(self):
         self.vim.command("JCcacheClear")
         classpath = self.vim.eval("g:JavaComplete_BaseDir") + "/javacomplete2/classpath/" + self.vim.eval("g:JavaComplete_ProjectKey")
-        os.remove(classpath)
+        if os.path.isfile(classpath):
+            os.remove(classpath)
         self.vim.command("call javacomplete#classpath#classpath#BuildClassPath()")
 
     def launchEmulator(self, emulator):
